Source code generator review. (#9)
* * Use the .NET 7 API ForAttributeWithMetadataName

* * Refactor complex lambda into separate function

* * Simplified the `IsSyntaxTargetForGeneration`
removed some unnecessary checks because we start use ForAttributeWithMetadataName method.

* code fix

* * Refactor GetTypeDeclarationForSourceGen method

* * Improved GetTypeKind method

* * Consider using RegisterImplementationSourceOutput instead of RegisterSourceOutput

* * Refactor method to get attribute typed argument name.
Simplified implementation and starting using full type name

* * test's fix

* * Refactor code generation in AutoVisitableGen

The GenerateCode method in AutoVisitableIncrementalSourceGenerator.cs has been refactored to directly use the namespace of the target symbol instead of getting the semantic model from the compilation. This simplifies the method, reducing unnecessary steps in the code generation process. The changes also involve a minor adjustment in the TypeDeclarationSyntax and renaming it with TypeNamespaceName.

* * Be careful using CompilationProvider

* * version update

* * code fixes

* * pr fix

* * test's